+++ packages/plugins/other/visitor-plugin-common/src/base-visitor.ts
@@ -46,7 +46,7 @@ export interface RawConfig {
    */
   scalars?: ScalarsMap;
   /**
-   * @default pascal-case#pascalCase
+   * @default change-case-all#pascalCase
    * @description Allow you to override the naming convention of the output.
    * You can either override all namings, or specify an object with specific custom naming convention per output.
    * The format of the converter must be a valid `module#method`.
@@ -55,19 +55,22 @@ export interface RawConfig {
    * Additionally you can set `transformUnderscore` to `true` if you want to override the default behavior,
    * which is to preserves underscores.
    *
+   * Available case functions in `change-case-all` are `camelCase`, `capitalCase`, `constantCase`, `dotCase`, `headerCase`, `noCase`, `paramCase`, `pascalCase`, `pathCase`, `sentenceCase`, `snakeCase`, `lowerCase`, `localeLowerCase`, `lowerCaseFirst`, `spongeCase`, `titleCase`, `upperCase`, `localeUpperCase` and `upperCaseFirst`
+   * [See more](https://github.com/btxtiger/change-case-all)
+   *
    * @exampleMarkdown
    * ## Override All Names
    * ```yml
    * config:
-   *   namingConvention: lower-case#lowerCase
+   *   namingConvention: change-case-all#lowerCase
    * ```
    *
    * ## Upper-case enum values
    * ```yml
    * config:
    *   namingConvention:
-   *     typeNames: pascal-case#pascalCase
-   *     enumValues: upper-case#upperCase
+   *     typeNames: change-case-all#pascalCase
+   *     enumValues: change-case-all#upperCase
    * ```
    *
    * ## Keep names as is
@@ -80,7 +83,7 @@ export interface RawConfig {
    * ```yml
    * config:
    *   namingConvention:
-   *     typeNames: pascal-case#pascalCase
+   *     typeNames: change-case-all#pascalCase
    *     transformUnderscore: true
    * ```
    */