+++ packages/plugins/other/visitor-plugin-common/src/naming.ts
@@ -1,7 +1,7 @@
 import { ASTNode } from 'graphql';
 import { resolveExternalModuleAndFn } from '@graphql-codegen/plugin-helpers';
 import { NamingConventionMap, NamingConvention, ConvertFn, ConvertOptions } from './types';
-import { pascalCase } from 'pascal-case';
+import { pascalCase } from 'change-case-all';
 import { convertNameParts, getConfigValue } from './utils';
 
 function getKind(node: ASTNode | string): keyof NamingConventionMap {