+++ packages/plugins/typescript/graphql-request/tests/graphql-request.spec.ts
@@ -55,7 +55,7 @@ describe('graphql-request', () => {
       usage,
     ]);
 
-    await validateTs(m);
+    validateTs(m);
 
     return m;
   };
@@ -87,7 +87,7 @@ async function test() {
 }`;
       const output = await validate(result, config, docs, schema, usage);
 
-      expect(result.content).toContain(`(print(FeedDocument), variables, requestHeaders));`);
+      expect(result.content).toContain(`(FeedDocument, variables, requestHeaders));`);
       expect(output).toMatchSnapshot();
     });
 
@@ -165,7 +165,8 @@ async function test() {
 
       const usage = `
 async function test() {
-  const client = new GraphQLClient('');
+  const Client = require('graphql-request').GraphQLClient;
+  const client = new Client('');
   const sdk = getSdk(client);
   
   await sdk.feed();
@@ -215,9 +216,9 @@ async function test() {
       const output = await validate(result, config, docs, schema, '');
 
       expect(output).toContain(`import * as Operations from './operations';`);
-      expect(output).toContain(`(print(Operations.FeedDocument), variables, requestHeaders));`);
-      expect(output).toContain(`(print(Operations.Feed2Document), variables, requestHeaders));`);
-      expect(output).toContain(`(print(Operations.Feed3Document), variables, requestHeaders));`);
+      expect(output).toContain(`(Operations.FeedDocument, variables, requestHeaders));`);
+      expect(output).toContain(`(Operations.Feed2Document, variables, requestHeaders));`);
+      expect(output).toContain(`(Operations.Feed3Document, variables, requestHeaders));`);
     });
   });
 });