+++ packages/plugins/typescript/resolvers/tests/ts-resolvers.spec.ts
@@ -136,6 +136,19 @@ describe('TypeScript Resolvers Plugin', () => {
   });
 
   describe('Config', () => {
+    it('onlyResolveTypeForInterfaces - should allow to have only resolveType for interfaces', async () => {
+      const config = {
+        onlyResolveTypeForInterfaces: true,
+      };
+      const result = await plugin(schema, [], config, { outputFile: '' });
+      const content = await validate(result, config, schema);
+
+      expect(content).toBeSimilarStringTo(`
+      export type NodeResolvers<ContextType = any, ParentType extends ResolversParentTypes['Node'] = ResolversParentTypes['Node']> = {
+        __resolveType: TypeResolveFn<'SomeNode', ParentType, ContextType>;
+      };`);
+    });
+
     it('optionalInfoArgument - should allow to have optional info argument', async () => {
       const config = {
         noSchemaStitching: true,
@@ -538,7 +551,6 @@ describe('TypeScript Resolvers Plugin', () => {
     expect(mergedOutputs).toContain(
       `something?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<QuerySomethingArgs, 'arg'>>;`
     );
-    validate(mergedOutputs);
   });
 
   it('Test for enum usage in resolvers (to verify compatibility with enumValues)', async () => {