+++ packages/plugins/typescript/graphql-request/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @graphql-codegen/typescript-graphql-request
 
+## 3.1.0
+
+### Minor Changes
+
+- af6fb509: Removed the unnecessary call to the print function, since graphql-request would call this function internally if needed.
+
+### Patch Changes
+
+- Updated dependencies [e947f8e3]
+- Updated dependencies [29b75b1e]
+- Updated dependencies [d4942d04]
+- Updated dependencies [1f6f3db6]
+- Updated dependencies [29b75b1e]
+  - @graphql-codegen/visitor-plugin-common@1.19.0
+  - @graphql-codegen/plugin-helpers@1.18.3
+
 ## 3.0.2
 
 ### Patch Changes