+++ packages/plugins/typescript/vue-apollo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @graphql-codegen/typescript-vue-apollo
 
+## 2.3.2
+
+### Patch Changes
+
+- 29b75b1e: enhance(namingConvention): use change-case-all instead of individual packages for naming convention
+- Updated dependencies [e947f8e3]
+- Updated dependencies [29b75b1e]
+- Updated dependencies [d4942d04]
+- Updated dependencies [1f6f3db6]
+- Updated dependencies [29b75b1e]
+  - @graphql-codegen/visitor-plugin-common@1.19.0
+  - @graphql-codegen/plugin-helpers@1.18.3
+
 ## 2.3.1
 
 ### Patch Changes